# Use valueOf instead of reflection for base types