Use `valueOf` at least for some base types allows to speed up some cases about 10%
before
```
Benchmark                            Mode  Cnt     Score    Error   Units
JmhTest._bothifyExpression          thrpt    5  1000.702 ±  9.367  ops/ms
JmhTest._letterifyExpression        thrpt    5   983.098 ± 18.326  ops/ms
JmhTest._optionsExpression          thrpt    5   675.234 ±  7.525  ops/ms
JmhTest._regexifyExpression         thrpt    5   310.129 ±  6.492  ops/ms
JmhTest.firstName                   thrpt    5   715.902 ± 13.890  ops/ms
JmhTest.initFaker                   thrpt    5   256.209 ±  0.784  ops/ms
JmhTest.oneMethodCallExpression     thrpt    5   595.780 ±  5.652  ops/ms
JmhTest.threeMethodsCallExpression  thrpt    5   166.082 ±  1.000  ops/ms

```

after
```
Benchmark                            Mode  Cnt     Score    Error   Units
JmhTest._bothifyExpression          thrpt    5  1067.565 ± 19.236  ops/ms
JmhTest._letterifyExpression        thrpt    5  1084.404 ±  3.447  ops/ms
JmhTest._optionsExpression          thrpt    5   676.119 ±  5.613  ops/ms
JmhTest._regexifyExpression         thrpt    5   321.421 ±  4.367  ops/ms
JmhTest.firstName                   thrpt    5   749.823 ±  6.373  ops/ms
JmhTest.initFaker                   thrpt    5   260.371 ±  1.357  ops/ms
JmhTest.oneMethodCallExpression     thrpt    5   630.570 ±  3.218  ops/ms
JmhTest.threeMethodsCallExpression  thrpt    5   165.601 ±  2.339  ops/ms
```